requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
Events==0.5
grpcio==1.78.0
idna==3.11
kafka-python==2.3.0
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
protobuf==7.34.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
